Adding time remaining countdown localization
